created components for PriceDisplay, HeaderButtons, Currency display, imported in Header component
